refactor: rename accountcmdutil to svcaccountcmdutil
