+++ pkg/cmd/serviceaccount/svcaccountcmdutil/svcaccount_util.go
@@ -1,0 +1,7 @@
+package svcaccountcmdutil
+
+import "github.com/redhat-developer/app-services-cli/pkg/cmd/serviceaccount/svcaccountcmdutil/credentials"
+
+var (
+	CredentialsOutputFormats = []string{credentials.EnvFormat, credentials.JSONFormat, credentials.PropertiesFormat}
+)
